Version 2 of the ROS robot PF
This robot is now capable of traversing nodes by calculating whether or not there is room to move calculating a trajectory and moving to the position

+++ src/tea/scripts/Robo_control.h
@@ -34,8 +34,8 @@ class TeaRobot
     std::vector<NAV_manager*>NavObjs;
     std::vector<Movement*>MoveObjs;
     //members
-    float count= 0;
-    float count1 =0;
+    int count= 0;
+    int count1 =0;
     //Nodes
     std::vector<std::vector<float>>Nodes;
     //Scan Data
